Lock to LTS node version
OK, so previously using the latest node version (17.4.0) we got lots of
"Unsupported engine" errors, one of which was to do with a package
called jest-watch-typeahead, and indeed, running the tests failed.

My theory is that the latest node isn't necessarily supported by
everything, so it's better to be on the stable (?) version instead:
https://nodejs.org/en/download/ says that the LTS version (16.13.2) is
"Recommended for most users" so let's go with that.