Use the same words, even after refreshing the page
A common annoyance with the previous system was that accidentally
refreshing the page resulted in a new word list being fetched.

Store the word list in localStorage so that even closing the browser
window and opening the app again gets the same word list (as distinct to
sessionStorage, which would allow multiple tabs with different cards).