Refactor: move state up to the app level
So up until now, cells have managed their own state (ie. whether they're
stamped or not), but that's not really appropriate.

We can see this in the fact that if we get a new card with a
stamped cell which (randomly) happens to be in the same position as it
was previously, then it will still be stamped in the new card, whereas
the same word in a different cell would not be stamped.

The solution is to have the state of _all_ cells managed by the App, and
give each cell a function with which it can update the stamped state.

The sequence of events is (as far as I understand it:

1. Cicking on a cell calls the `setStamped` callback function.
2. This updates the state of that cell on the `App` component
3. Since `cellPropsList` depends on the state, and the render in `App`
   depends on `cellPropsList`, this triggers a re-render
4. The cell which was clicked gets a new value for `stamped`, so it
   re-renders, and sets the stamped class