Store all the CellData state in the session
This actually simplifies things somewhat: now that we're not updating
the words independently of the stamps, our function for building a new
state (now `newCellDataList`) can always use a new set of words, since
we only build a new state if we didn't find one in the session.

This comes at the cost of introducing a sync problem: we want to make
sure that our component state (created with `setState`) is always in
sync with `localStorage`. We make this easier by adding a
setAndSaveCellDataList function which sets the state in both places.