Refactor: extract behaviour to session objects
The observation here is that we're using a generic session concept which
doesn't have anything to do with our actual data. That insight allows us
to create a couple of classes with generic types.

Side note: actually there's a restriction here: we can only pass objects
which can be serialised to JSON. But in practice we're only actually
using one type, which is definitely serialisable, so the concern is
academic.

- JsonSession serializes arbitrary data in localStorage as JSON - it
  might return null
- GenericJsonSession serializes arbitrary data in JsonSession, and
  accepts an init function which can be used to initialise the session
  with new data, so that it always returns data.

Note that previously we weren't actually using a setter on the session.
Now that we are, typescript points out that the value of the setter
needs to match the return value of the getter. Since the getter can
return null, the setter needs to accept null.

We're not expecting to ever send null to the setter, but let's make the
behaviour sensible (why not?): rather than allowing it to set "null" as
a string, we can remove the value from localStorage.

ALSO extract the words function to its own file - if only to make it
clear that the imported functions only relate to that one function.