Fix failing test of New card button
This was failing since we changed the key to use the word rather than
the column index. This failed because `cells` was still referring to the
old component, which is no longer visible, having been replaced with a
new component with a different word.

We fix this by fetching the cell array again.