Github action: run tests and checks
This adds a github action which will run the server tests on all PRs and
any time there's a push to main.

Use `npm ci` instead of `npm install` for a faster build (?) and to
ensure that package-lock.json isn't changed when installing packages.

I'm not super sure if the typecheck is necessary (will we get
compilation errors when running the app if there are type errors?)