Lint for accessibility issues
The a11y plugin lints for accessibility issues.

We had one issue:

> error  Non-interactive elements should not be assigned mouse or keyboard event listeners
> jsx-a11y/no-noninteractive-element-interactions

This is explained here:

  https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/main/docs/rules/no-noninteractive-element-interactions.md

The issue is that we've attached a click event to a table cell and table
cells aren't expected to be interactive elements.

The solution we chose is to use the aria roles "grid" and "gridcell" to
communicate that this isn't a table storing data - instead it's used for
layout:

  https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/main/docs/rules/no-noninteractive-element-interactions.md#option-2-convert-the-table-into-an-aria-grid
  https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Grid_Role
  https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/gridcell_role