Add a Wordle-style "share" version of the board
Map all the stamps to emoji, and copy to the clipboard.

Testing this involves a bit of a hack, but I feel OK about it: basically
the clipboard is (for sensible reasons!) secured, requiring the user to
confirm that an app is allowed to read from it.

But it's perfectly OK to write to it and (apparently?) paste from it.
Actually The testing library uses some kind of stub replacing the
clipboard in order to get around this permissions issue.

So we check the contents of the clipboard by rendering a text area that
we can paste it into.

Note that it's generally awkward to create and compare a multiline
string in code, hence the use of stripIndent and trim() in both the test
and the code. This is not ideal, but I feel OK about it.

One more thing: the OSX white square emoji isn't the same as the one
used on Discord, so it displays as a different size to the blue square,
and with different corners. I grabbed the white emoji from here instead:

  https://emojidb.org/square-emojis