Allow soft-hyphens in words
Some words are a bit too long to fit into a square, and by default they
break in places that aren't so sensible.

The solution is to use soft-hypens: a character which is invisible when
there's enough space, but defines where a word will break, and shows a
hyphen where it breaks.

The challenge is that (to prevent xss attacks), React quite sensibly
displays it as a string rather than evaluating it.

The solution to that is to replace it with the unicode version.

Yes, we could just put the unicode version directly into the word list,
but it looks more broken and it's much less obvious what it means
(though \uxxx does get syntax highlighted while &shy; doesn't, so it's
tradeoffs all the way down)

Use a regular expression with the global modifier (g) to make sure that
if we ever have multiple soft hyphens then they are all replaced.