Add descriptions to words in the word list
Using parentElement in the test because the testing library can only
get text using the whole text of nodes, not including child nodes, so we
instead look at the text content of the enclosing element.

I'm not sure about this approach where we're passing one component as a
prop to another one in order to wrap the text - this is the only way I
know how to do conditional rendering in React.