Tweak the share button
1. Always copy to clipboard, even if share functionality is available

This is because for some share types (eg. Facebook) the "text" part is
discarded, and I want people to be able to share their cards, not just
the links.

2. Update title

When selecting "copy" from the share menu, the "text" is presented as a
link, with the "url" as the URL, so the text should reflect what the
link is.