Choose language from a language picker
Need to ass some options:

supportedLngs

This is needed to allow i18next to pick the best match out of
all the browser's supported languages list (navigator.languages).
Otherwise only the first language in the browwer's list is used,
regardless of whether it matches any supported languages or if there are
any.

Export this to try and make sure that the picker and the i18n config
stay in sync.

nonExplicitSupportedLngs

This is needed to eg. allow fr-BE to be treated as fr. Seems like it's
disabled by default because it doesn't play nice with loading
translations from external sources.