Styling of the card
The font spreading is hacky, but seems to be the only way to get a good
responsive behavour?

Use an off-white and off-black colour for the card, and black and white
for the rest.

Add a pixel to the dividing lines to make it feel more bingo, but then
we need to very very slightly make the font size smaller to compensate.