Add a noise filter
Make it look less crisp and more like an old-school bingo card.

This was EXTREMELY fiddly because the filter makes the dark parts
lighter and the light parts darker, so we need to compensate by making
the light and dark colours more extreme.

Also the QR code was a black SVG - I couldn't get this to work with CSS
variables so just hard-coded the colour.