Add a noise overlay
Make it look less crisp and more like an old-school bingo card.

This feels a little hacky, but adding the noise filter with CSS was
problematic: it needed elements to have semi-transparent backgrounds,
but then components stacked on top of each other would have different
transparencies and different levels of noise.

It was also a pain to compensate for the opacity by making colours
darker and lighter.