Fix: QR code button on iOS
The issue seems to be with font rendering - on iOS there was extra space
below the QR code button, which is supposed to be square.

Two fixes here, maybe only one is necessary?

- Use a background image instead of an img tag
- set line height to 1

I guess line height is usually a bit more than 1 for legibility of
multiple lines of text, but for text in a button it's not needed?