Fix single page app routing with spa-github-pages
https://github.com/rafgraph/spa-github-pages/

So it turns out GitHub Pages doesn't actually do dynamic routing: if we
try to access https://dgmstuart.github.io/bingo-frontend/numbers, then
github looks for an html document matching that name, and doesn't find
one so returns a 404.

This was confusing for a while because it seemed like those pages were
working on desktop and not mobile, but actually it seems like it was
just caching which allowed them to be displayed?

This approach is to define a specific 404 page which routes back to the
index which then handles the route.