Separate session data per config
Pass the file name to the session data key, so that we get separate
entries per config.

This has the effect of auto-switching the grid when choosing a new
config - that's not actually an expected use case but it feels correct
that the app should only be able to show a consistent set of config and
words.

Jump through some hoops to avoid not found resulting in junk
localstorage entries:

Since we fall back to a default list if the name is not recognised
(maybe a bad idea?), the param doesn't always match the list filename,
and we don't want eg http://localhost:5173/bingo-frontend/foo to result
in storing bingoSession-foo in localstorage.

Maybe we can skip all that nonsense and do some kind of redirect or not
found instead?