Redo approach to button active states
This didn't actually work: it assumed that there were styles set on the
element with `style=""` (?) but the styling is actually done entirely
with CSS.

In some ways the setStyles approach is maybe more elegant? But it means
we can't use the CSS variables (?), and overall an "active" class just
seems simpler.