Move video list URL to config
The video list is a property of the game - eg. if we did Balboa bingo it
would be a different list.

This is getting a bit out of hand: maybe the better approach is to use a
provider - that makes config feel like global state, but maybe that's
not wrong?