Configure site to run with a custom domain
Now that I've bought a domain, I guess it needs to run on the root?
Prevously it was running on

    https://dgmstuart.github.io/bingo-frontend/

...and now it's just

    https://teamlindybingo.com/

ROUTING:

To prevent infinite rewrites we need to set pathSegmentsToKeep back to
0, otherwise we get infinite ~and~ being added:

    https://teamlindybingo.com/balboa/?/&/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/~and~/

  https://github.com/rafgraph/spa-github-pages/blob/v6.1.0/404.html#L18-L24

CNAME:

When adding the custom domain in the Github UI, it creates a commit
adding a CNAME file with the domain. But it adds it to the gh-pages
branch, so every time I build the app, it disappears and we lose the
custom domain.

By putting this file in /public, it gets generated in the final
distribution, so the custom domain persists.